pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doppelganger-tracker"
version = "1.0.0"
description = "Academic research tool for disinformation analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Doppelganger Tracker Team"}
]
keywords = [
    "disinformation",
    "analysis",
    "nlp",
    "propaganda",
    "research",
    "telegram",
    "fact-checking"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "feedparser>=6.0.0",
    "telethon>=1.34.0",
    "spacy>=3.7.0",
    "langdetect>=1.0.9",
    "networkx>=3.2.0",
    "python-louvain>=0.16",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "plotly>=5.18.0",
    "streamlit>=1.29.0",
    "apscheduler>=3.10.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "ipython>=8.18.0",
]
ml = [
    "scikit-learn>=1.3.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "sentence-transformers>=2.2.0",
]

[project.scripts]
doppelganger = "main:main"

[project.urls]
Homepage = "https://github.com/your-repo/doppelganger-tracker"
Documentation = "https://github.com/your-repo/doppelganger-tracker#readme"
Repository = "https://github.com/your-repo/doppelganger-tracker"
Issues = "https://github.com/your-repo/doppelganger-tracker/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "database*", "collectors*", "analyzers*", "dashboard*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
known_first_party = ["config", "database", "collectors", "analyzers", "dashboard"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "tests/",
    "migrations/",
    "venv/",
]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "venv",
    "migrations",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "*/migrations/*",
    "conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
